+++ CausalForce/train_causal.py
@@ -110,7 +110,7 @@ class CausalForce(pl.LightningModule):
             elif len(gt_risk_ids) == 0:
                 gt_zeros = torch.zeros_like(pred_risk_score_H8[:N])
                 loss_scoreH8_i = F.binary_cross_entropy(
-                    pred_risk_score_H8[:N], gt_zeros, reduction='mean')
+                    pred_risk_score_H8[:N].float(), gt_zeros.float(), reduction='mean')
                 loss_htsc_i = htsc_loss(
                     hx_seq=outputs["hx_seq"][i],
                     pred_risk_type=outputs["risk_type"][i],
@@ -156,7 +156,7 @@ class CausalForce(pl.LightningModule):
                 if len(matched_pred) == 0:
                     gt_zeros = torch.zeros_like(pred_risk_score_H8[:N])
                     loss_scoreH8_i = F.binary_cross_entropy(
-                        pred_risk_score_H8[:N], gt_zeros, reduction='mean')
+                        pred_risk_score_H8[:N].float(), gt_zeros.float(), reduction='mean')
                     loss_cf_i = counterfactual_loss(
                         pred_direct[:N], pred_indirect[:N], gt_zeros)
 
@@ -181,7 +181,7 @@ class CausalForce(pl.LightningModule):
 
                     if len(non_risk_pred) == 0:
                         loss_scoreH8_i = F.binary_cross_entropy(
-                            preds, gts, reduction='mean')
+                            preds.float(), gts.float(), reduction='mean')
                         loss_cf_i = counterfactual_loss(
                             d_stack, i_stack, gts)
                     else:
@@ -191,8 +191,8 @@ class CausalForce(pl.LightningModule):
                         nr_i = torch.stack(non_risk_indirect)
 
                         loss_scoreH8_i = (
-                            F.binary_cross_entropy(preds, gts, reduction='mean')
-                            + F.binary_cross_entropy(nr_preds, nr_gt, reduction='mean'))
+                            F.binary_cross_entropy(preds.float(), gts.float(), reduction='mean')
+                            + F.binary_cross_entropy(nr_preds.float(), nr_gt.float(), reduction='mean'))
                         loss_cf_i = (
                             counterfactual_loss(d_stack, i_stack, gts)
                             + counterfactual_loss(nr_d, nr_i, nr_gt))
